Added LIBMTP_Toggle_Errorstack to disable (or enable) the using the device errorstack
In a lot of situations, the client application doesn't care about the entries
added to the errorstack each time that an error is encountered by the device.
But, even in that case, the client application has the burden to clear the
errorstack regularly.
This new function allows to disable that behavior after having opened a device.